Major overhaul of "modifiedLM" connection method w/ more to come... Now that data structure is in place for passing CDInfo, have collision detection actually return the index of the multibody that robot cfg is first found to be in collision with.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@96 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ OBPRM/ConnectMapNodes.h
@@ -135,25 +135,25 @@ public:
   int 	 GetIterations () const;
   int 	 GetStepFactor () const;
   int    GetMaxNum     () const;
-  int    GetRFactor    () const;
+  double GetRFactor    () const;
 
 protected:
   //===================================================================
   // Data
   //===================================================================
-  char  name[80];
-  CNF   connector;          // ptr to connection strategy code
-  EID   cnid;
-  int 	numEdges;          // used by random algm 
-  int   kclosest;          // used by closest algm & modifiedLM
-  int   maxNum;            // used by modifiedLM
-  int   rfactor;           // used by modifiedLM
-  int 	kpairs;            // used by components algm
-  int 	smallcc;           // used by components & rrt algm
-  int 	stepFactor;        // used by rrt algm
-  int 	iterations;        // used by rrt algm
-  int 	k_other;           // used by obstBased algm
-  int 	k_self;            // used by obstBased algm
+  char   name[80];
+  CNF    connector;          // ptr to connection strategy code
+  EID    cnid;
+  int 	 numEdges;          // used by random algm 
+  int    kclosest;          // used by closest algm & modifiedLM
+  int    maxNum;            // used by modifiedLM
+  double rfactor;           // used by modifiedLM
+  int 	 kpairs;            // used by components algm
+  int 	 smallcc;           // used by components & rrt algm
+  int 	 stepFactor;        // used by rrt algm
+  int 	 iterations;        // used by rrt algm
+  int 	 k_other;           // used by obstBased algm
+  int 	 k_self;            // used by obstBased algm
  
 private:
 };
